pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Mike Kolotovich <k2punk@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
gunicorn = "^21.2.0"
django-bootstrap5 = "^23.4"
django-filter = "^24.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
